pyproject.toml
[project]
name = "final-fantasy-eventbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "김동환", email = "kimdonghwandeveloper@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "openai>=2.31.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]

[project.scripts]
final-fantasy-eventbot = "final_fantasy_eventbot:main"

[build-system]
requires = ["uv_build>=0.9.18,<0.10.0"]
build-backend = "uv_build"
